+++ src/examples/eventbrite/src/backend/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './event';
+export * from './order';
